+++ ev.h
@@ -35,7 +35,7 @@
  * 
  * ### Bug Fixes
  * 1. fix: loop may wait infinite with EV_LOOP_MODE_ONCE if there are only endgame events
- * 
+ * 2. fix: memory leak in create process
  * 
  * ## v0.0.8 (2023/04/07)
  * 
@@ -204,7 +204,7 @@ extern "C" {
 /**
  * @brief Development version.
  */
-#define EV_VERSION_PREREL           3
+#define EV_VERSION_PREREL           4
 
 /**
  * @brief Version calculate helper macro.